ai-agents/requirements.txt
# Core CrewAI and FastAPI
crewai
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# AI and Language Models
openai==1.13.3
# langchain packages managed by crewai
transformers==4.36.0
torch==2.1.0

# Speech Processing
speechrecognition==3.10.0
gtts==2.4.0
pydub==0.25.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1

# Web and API
requests==2.31.0
httpx==0.25.2
websockets==12.0
aiofiles==23.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0

# Database and Storage
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
